backend/requirements.txt
llama-index
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
python-dotenv
chromadb
fastapi
uvicorn